backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
llama-cpp-python==0.2.85
PyMuPDF==1.24.9
beautifulsoup4==4.12.3
requests==2.32.3
nltk==3.9.1
pymongo==4.8.0
sentence-transformers==3.0.1